Set symfony/yaml version to '~2.4.0' in composer.json
This is a more appropriate/reliable practice for getting the latest
patches on the 2.4 branch.

For descriptions of the version contstraint operators, see:

    https://getcomposer.org/doc/articles/versions.md

The `.x-dev` notation is intended for use in defining branch-aliases
and does not usually correspond to an actual release tag. The branch
alias is used to allow a library's branch name to satisfy a version
constraint within a project that requires it. See:

    https://getcomposer.org/doc/articles/aliases.md#branch-alias